refactor the type files to separate file
